Move buffer state tracking into a module, add unit tests
